feat(ci): Add release workflow
This adds a workflow to build the Python package when a tag matching `feature_artifacts-helper_*` is created. It will upload the build artifacts (`tar.gz` sdist and `.whl` bdist).

There is also the option of automatically creating a release and
attaching the build artifacts to it. See
eebf0a5970533761dd94f764a3abfd1d24faecb3 and
https://github.com/delilahw/codespace-features/releases/tag/feature_artifacts-helper_1.1-alpha
for an example.