chore: Update ruff `src`s to fix detection of first-party modules in isort
Previously, when we imported our own module
`codespaces_artifacts_helper_keyring` in tests, it was not being
detected by ruff (linter) as a first-party module. This affected the
linter's ability to accurately separate import blocks.

We fix the issue in this commit by including the `/test` directory in
ruff's `src`s. Now, the linter will put our own modules in a separate
first-party import block, as expected.